[30/12/2012] Version 0.4.171.890a Adding functionalities: - review sqlite database struct - review global state tool - review count and last run bug on prefetch - review registry editor bug of date view - reviex registry editor bug in view of deleted registry key and value ! - review extract of SID owner on registry key - review IE history read - Enable command line of RtCA - add menu entry for reload session list - adding in copy of log file : usb and firewall file history - add file copy with VSS
Tempory disable:
- no x64 binary update

+++ RTCA/tests/test_registry_users.c
@@ -197,7 +197,7 @@ void Scan_registry_user_local(sqlite3 *db, unsigned int session_id)
                   case 0:snprintf(type,MAX_PATH,"%lu : %s",pBuf_info->usri2_priv,cps[TXT_MSG_GUEST].c);break;
                   case 1:snprintf(type,MAX_PATH,"%lu : %s",pBuf_info->usri2_priv,cps[TXT_MSG_USER].c);break;
                   case 2:snprintf(type,MAX_PATH,"%lu : %s",pBuf_info->usri2_priv,cps[TXT_MSG_ADMIN].c);break;
-                  default:snprintf(type,MAX_PATH,"0x%02X : %s",(pBuf_info->usri2_priv & 0xff),cps[TXT_MSG_UNK].c);break;
+                  default:snprintf(type,MAX_PATH,"0x%02X : %s",(unsigned int)(pBuf_info->usri2_priv & 0xff),cps[TXT_MSG_UNK].c);break;
                 }
 
                 addRegistryUsertoDB("NETAPI32",name, RID, SID, group, description,
